backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic-settings
python-multipart
meilisearch
python-jose[cryptography]
passlib[bcrypt]
httpx
pytest
pytest-asyncio
aiofiles
